backend/rag/requirements.txt
langchain
chromadb
sentence-transformers
flask
flask_cors
torch
pyMySql[rsa]
peewee
python-dotenv
redis
jwt
PyJWT
waitress